requirements.txt
django==1.7
django-autoslug
django-ckeditor
django-positions
django-admin-sortable2
Pillow
